tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssKeywordValue.html

<!doctype html>
<meta charset="utf-8">
<title>CSSKeywordValue</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#csskeywordvalue">
<meta name="assert" content="Test CSSKeywordValue constructor and attributes" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log">
<script>
'use strict';

const gTestArguments = [
  { keyword: 'initial', description: 'a CSS wide keyword' },
  { keyword: 'auto', description: 'a CSS keyword' },
  { keyword: 'lemon', description: 'an unsupported CSS keyword' },
  { keyword: '3! + 4@', description: 'a string containing multiple tokens' },
  { keyword: '☺', description: 'a unicode string' },
];

for (const {keyword, description} of gTestArguments) {
  test(() => {
    const result = new CSSKeywordValue(keyword);

    assert_not_equals(result, null,
        'A CSSKeywordValue should be created');
    assert_equals(result.value, keyword,
        'Value attribute should be same as passed in the constructor');
  }, 'CSSKeywordValue can be constructed from ' + description);

  test(() => {
    let result = new CSSKeywordValue('auto');
    result.value = keyword;

    assert_equals(result.value, keyword,
        'Value attribute should be same as passed in the setter');
  }, 'CSSKeywordValue.value can be updated to ' + description);
}

</script>
